requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-auth-ldap==1.2.15
django-constance==2.0.0
django-cors-headers==2.1.0
django-filter==2.0.0
django-guardian==1.4.9
django-picklefield==1.0.0
django-suit==0.2.26
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
PyJWT==1.5.3
pyldap==2.4.37
pytz==2018.7
requests==2.18.4
six==1.10.0
urllib3==1.22
websocket-client==0.44.0
uwsgi
psycopg2

